fix(products): wrap products page with Suspense and move useSearchParams logic into ProductsClient
